# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/auth/otp/request": {
      "post": {
        "operationId": "AuthController_requestOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Запросить OTP (НФ-05: TTL 2 мин, resend 45 с)",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/otp/verify": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Проверить OTP → tokens + is_new_user",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Обновить пару токенов (ротация refresh)",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/me": {
      "get": {
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Текущий пользователь + профиль специалиста",
        "tags": [
          "me"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/role": {
      "post": {
        "operationId": "UsersController_switchRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Переключить/активировать роль (Р-01…Р-05)",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/devices": {
      "post": {
        "operationId": "UsersController_registerDevice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/v1/specialist/profile": {
      "post": {
        "operationId": "UsersController_specialistProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecialistProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Анкета специалиста (S-05)",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/categories": {
      "get": {
        "operationId": "CategoriesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Справочник одобренных категорий (К-06)",
        "tags": [
          "categories"
        ]
      }
    },
    "/v1/categories/suggest": {
      "post": {
        "operationId": "CategoriesController_suggest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Предложить категорию (К-02 → pending)",
        "tags": [
          "categories"
        ]
      }
    },
    "/v1/specialist/verification": {
      "post": {
        "operationId": "VerificationController_submitVerification",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "S-06: селфи + селфи с документом (В-06)",
        "tags": [
          "specialist"
        ]
      }
    },
    "/v1/specialist/diploma": {
      "post": {
        "operationId": "VerificationController_submitDiploma",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "S-05a: диплом jpg/png/pdf ≤10МБ (ДС-*, НФ-09)",
        "tags": [
          "specialist"
        ]
      }
    },
    "/v1/orders": {
      "post": {
        "operationId": "OrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "S-20: создать заказ",
        "tags": [
          "orders"
        ]
      }
    },
    "/v1/orders/my": {
      "get": {
        "operationId": "OrdersController_myOrders",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Мои заказы (заказчик, ИЗ-02)",
        "tags": [
          "orders"
        ]
      }
    },
    "/v1/orders/feed": {
      "get": {
        "operationId": "OrdersController_feed",
        "parameters": [
          {
            "name": "lat",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "S-11: лента специалиста (гео + фильтры Ф-07 + блок «Новые»)",
        "tags": [
          "orders"
        ]
      }
    },
    "/v1/orders/{id}": {
      "get": {
        "operationId": "OrdersController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "S-12: карточка заказа",
        "tags": [
          "orders"
        ]
      }
    },
    "/v1/orders/{id}/hide": {
      "post": {
        "operationId": "OrdersController_hide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "S-11: скрыть заказ (свайп влево)",
        "tags": [
          "orders"
        ]
      }
    },
    "/v1/orders/{id}/bids": {
      "post": {
        "operationId": "OrdersController_createBid",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBidDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "S-13: откликнуться на заказ",
        "tags": [
          "orders"
        ]
      },
      "get": {
        "operationId": "OrdersController_orderBids",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "S-23: отклики на заказ (заказчик)",
        "tags": [
          "orders"
        ]
      }
    },
    "/v1/bids/my": {
      "get": {
        "operationId": "BidsController_myBids",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "S-14: мои отклики (специалист)",
        "tags": [
          "bids"
        ]
      }
    },
    "/v1/bids/{id}/accept": {
      "post": {
        "operationId": "BidsController_accept",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "S-24: принять отклик (каскад «Не выбран» + чат + комиссия)",
        "tags": [
          "bids"
        ]
      }
    },
    "/v1/bids/{id}/decline": {
      "post": {
        "operationId": "BidsController_decline",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "S-24: отклонить отклик",
        "tags": [
          "bids"
        ]
      }
    }
  },
  "info": {
    "title": "Zovu API",
    "description": "C2C-маркетплейс услуг — REST /v1. Source of truth: docs/04-api.md",
    "version": "0.2.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RequestOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+77012345678",
            "description": "Номер КЗ (нормализуется на сервере)"
          }
        },
        "required": [
          "phone"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+77012345678"
          },
          "code": {
            "type": "string",
            "example": "1111",
            "description": "OTP из SMS (dev — всегда 1111)"
          }
        },
        "required": [
          "phone",
          "code"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ]
      },
      "UpdateMeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "dob": {
            "type": "string",
            "example": "1998-05-12"
          },
          "lang": {
            "type": "string",
            "enum": [
              "ru",
              "kk"
            ]
          }
        }
      },
      "SwitchRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "client",
              "specialist"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
      "RegisterDeviceDto": {
        "type": "object",
        "properties": {
          "fcmToken": {
            "type": "string"
          },
          "platform": {
            "type": "string",
            "example": "web"
          }
        },
        "required": [
          "fcmToken",
          "platform"
        ]
      },
      "SpecialistProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "dob": {
            "type": "string",
            "example": "1990-01-01"
          },
          "mainCategoryId": {
            "type": "string"
          },
          "categoryIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "about": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "mainCategoryId"
        ]
      },
      "SuggestCategoryDto": {
        "type": "object",
        "properties": {}
      },
      "OrderFiltersDto": {
        "type": "object",
        "properties": {
          "certifiedOnly": {
            "type": "boolean"
          },
          "minRating": {
            "type": "number"
          },
          "minOrders": {
            "type": "number"
          },
          "maxDistanceKm": {
            "type": "number"
          }
        }
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "photos": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "budget": {
            "type": "number"
          },
          "address": {
            "type": "string"
          },
          "lat": {
            "type": "number"
          },
          "lng": {
            "type": "number"
          },
          "filters": {
            "$ref": "#/components/schemas/OrderFiltersDto"
          }
        },
        "required": [
          "categoryId",
          "title",
          "description",
          "budget",
          "address"
        ]
      },
      "CreateBidDto": {
        "type": "object",
        "properties": {
          "price": {
            "type": "number",
            "description": "Цена специалиста (предзаполнена бюджетом заказа)"
          }
        },
        "required": [
          "price"
        ]
      }
    }
  }
}